Updated Alert component
